+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.0]
+
+- renamed `khao-ui-textfield` into `khao-text-field`
+- renamed `khao-ui-selectfield` into `khao-select-field`
+
 ## [3.1.6]
 
 - added `mortar-pestle` icon